Remove old conversion code and update printf. Start work on FILE and stdout. Start work on tty. Remove awful unneeded ring buffer code. Move char checking code to ctype.h. Unused likely not working smp code.

+++ kernel/src/stdlib.c
@@ -12,7 +12,7 @@ int atoi(const char* buf){
 	}
 	return output*sign;
 }
-int atol(const char* buf){
+long atol(const char* buf){
 	long output=0;
 	long index=0;
 	long sign =1;
